requirements.txt
Kivy==2.3.1
kivymd==2.0.1.dev0
ollama==0.5.3
psutil==7.0.0
PyYAML==6.0.2
pytest==8.3.5
requests==2.32.4
